# Report generation performance improvements