--- a/src/Reporting/Report.php
+++ b/src/Reporting/Report.php
@@ -303,7 +303,7 @@ class Report implements Arrayable, Jsonable
     public function save()
     {
         File::put($this->path(), YAML::dump($this->raw = [
-            'date' => $this->date ?? time(),
+            'date' => $this->date ?? now()->timestamp,
             'status' => $this->status(),
             'score' => $this->score(),
             'pages_crawled' => $this->pagesCrawled(),