… src/Reporting/Rules/UniqueMetaDescription.php
… src/Reporting/Rules/UniqueTitleTag.php